requirements.txt
# HTTP请求
requests==2.31.0

# Excel处理
openpyxl==3.1.2

# YAML处理
PyYAML==6.0.1

# 测试框架
pytest==7.4.3
pytest-html==4.1.1
allure-pytest==2.13.5

# 日志
loguru==0.7.2
